Final contract before going live
Added Gerd and optimized some things, reduced contract from 1430 bytes to around 1200 bytes, but contract creation still too expensive, we don't have enough in the pot to get through with 21gwei and smaller gas values get stuck. So the contract will remain offline and a simple program will check the same and send out the tx this way once we reach <12.5k or >25k on coinmarketcap.